requirements.txt
streamlit
torch
torchvision
opencv-python-headless
Pillow
numpy
scikit-image